requirements.txt
pytorch
torchvision
scikit-image
scipy
numpy
tqdm
